Watch.
I couldn't solve that bug, but now I'll add a system to check config file entries better.

Watch.

I have a `Class<?>` field in `AppConfigEntry` right now. This will now become a `Function<String, String>` instead. If there *is* a return value, we ask the user to put a value in again.

I considered using an exception. Just won't do it for now.